+++ filefinder/_filefinder.py
@@ -5,12 +5,18 @@ import logging
 import os
 import re
 import warnings
+from typing import Any
 
 import numpy as np
 import pandas as pd
 import parse
 
-from .utils import _find_keys, natural_keys, product_dict, update_dict_with_kwargs
+from filefinder._utils import (
+    _find_keys,
+    natural_keys,
+    product_dict,
+    update_dict_with_kwargs,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -22,12 +28,50 @@ keys: {repr_keys}
 """
 
 
+def _deprecate_allow_empty(**kwargs):
+
+    _allow_empty = kwargs.get("_allow_empty")
+
+    if _allow_empty is not None:
+        raise TypeError("`_allow_empty` has been deprecated in favour of `on_empty`")
+
+
+_RESERVED_PLACEHOLDERS = {"keys", "on_parse_error", "on_empty", "_allow_empty"}
+
+
+def _assert_valid_keys(keys) -> None:
+
+    for key in _RESERVED_PLACEHOLDERS:
+        if key in keys:
+            raise ValueError(f"'{key}' is not a valid placeholder")
+
+
+def _assert_unique(df) -> None:
+
+    duplicates = df.duplicated()
+
+    if duplicates.any():
+        duplicated = df[duplicates].head()
+        msg = f"Non-unique metadata detected.\nFirst five duplicates:\n{duplicated}"
+
+        raise ValueError(msg)
+
+
 class _FinderBase:
     def __init__(self, pattern, suffix=""):
 
         self.pattern = pattern
         self.keys = _find_keys(pattern)
+        _assert_valid_keys(self.keys)
         self.parser = parse.compile(self.pattern)
+
+        if self.parser.fixed_fields:
+            msg = (
+                "Only named fields are currently allowed: avoid empty braces and"
+                " leading underscores."
+            )
+            raise ValueError(msg)
+
         self._suffix = suffix
 
         # replace the fmt spec - add the capture group again
@@ -35,7 +79,7 @@ class _FinderBase:
             r"\{([A-Za-z0-9_]+)(:.*?)\}", r"{\1}", pattern
         )
 
-    def create_name(self, keys=None, **keys_kwargs):
+    def create_name(self, keys=None, **keys_kwargs) -> str:
         """build name from keys
 
         Parameters
@@ -62,8 +106,8 @@ class _Finder(_FinderBase):
         return cond_dict
 
     def find(
-        self, keys=None, *, on_parse_error="raise", _allow_empty=False, **keys_kwargs
-    ):
+        self, keys=None, *, on_parse_error="raise", on_empty="raise", **keys_kwargs
+    ) -> "FileContainer":
         """find files in the file system using the file and path (folder) pattern
 
         Parameters
@@ -74,8 +118,9 @@ class _Finder(_FinderBase):
         on_parse_error : "raise" | "warn" | "ignore", default: "raise"
             What to do if a path/file name cannot be parsed. If "raise" raises a ValueError,
             if "warn" raises a warning and if "ignore" ignores the file.
-        _allow_empty : bool, default: False
-            If False (default) raises an error if no files are found. If True returns
+        on_empty : "raise" | "warn" | "allow", default: "raise"
+            Behaviour when no files are found: "raise" (default) raises a ValueError,
+            "warn" raises a warning. For "warn" and "allow" an empty FileContainer is returned.
             an empty list.
         **keys_kwargs : {key: indexer, ...}, optional
             The keyword arguments form of ``keys``. When the same key is passed in
@@ -87,11 +132,18 @@ class _Finder(_FinderBase):
 
         """
 
+        _deprecate_allow_empty(**keys_kwargs)
+
         keys = update_dict_with_kwargs(keys, **keys_kwargs)
 
-        if on_parse_error not in ["raise", "warn", "ignore"]:
+        if on_parse_error not in ("raise", "warn", "ignore"):
             raise ValueError(
                 f"Unknown value for 'on_parse_error': '{on_parse_error}'. Must be one of 'raise', 'warn' or 'ignore'."
+            )
+
+        if on_empty not in ("raise", "warn", "allow"):
+            raise ValueError(
+                f"Unknown value for 'on_empty': '{on_empty}'. Must be one of 'raise', 'warn' or 'allow'."
             )
 
         # wrap strings and scalars in list
@@ -112,30 +164,57 @@ class _Finder(_FinderBase):
 
             all_patterns.append(full_pattern)
 
-        if all_paths:
-            df = self._parse_paths(all_paths, on_parse_error=on_parse_error)
-        elif _allow_empty:
-            return []
-        else:
+        if len(all_paths) == 0:
             msg = "Found no files matching criteria. Tried the following pattern(s):"
             msg += "".join(f"\n- '{pattern}'" for pattern in all_patterns)
-            raise ValueError(msg)
 
-        fc = FileContainer(df)
+            if on_empty == "raise":
+                raise ValueError(msg)
+            elif on_empty == "warn":
+                # TODO: correct stack level
+                warnings.warn(msg)
 
-        len_all = len(fc.df)
-        len_unique = len(fc.combine_by_key().unique())
+        # NOTE: also creates the correct (empty) df if no paths are found
+        df = self._parse_paths(all_paths, on_parse_error=on_parse_error)
+        _assert_unique(df)
 
-        if len_all != len_unique:
-            duplicated = fc.df[fc.df.duplicated()]
-            msg = f"This query leads to non-unique metadata. Please adjust your query.\nFirst five duplicates:\n{duplicated.head()}"
+        return FileContainer(df)
 
+    def find_single(self, keys=None, **keys_kwargs) -> "FileContainer":
+        """
+        find exactly one file/ path in the file system using the file and path pattern
+
+        Parameters
+        ----------
+        keys : dict
+            Dictionary containing keys to create the search pattern.
+        **keys_kwargs : {key: indexer, ...}, optional
+            The keyword arguments form of ``keys``. When the same key is passed in
+            ``keys`` and ``keys_kwargs`` the latter takes priority.
+
+        Notes
+        -----
+        Missing ``keys`` are replaced with ``"*"``.
+
+        Raises
+        ------
+        ValueError : if more or less than one file/ path is found
+        """
+
+        fc = self.find(keys, on_parse_error="raise", on_empty="raise", **keys_kwargs)
+
+        if len(fc) > 1:
+            n_found = len(fc)
+            msg = (
+                f"Found more than one ({n_found}) files/ paths. Please adjust your"
+                f" query.\nFirst five files/ paths:\n{fc.df.head()}"
+            )
             raise ValueError(msg)
 
         return fc
 
     @staticmethod
-    def _glob(pattern):
+    def _glob(pattern) -> list[str]:
         """Return a list of paths matching a pathname pattern
 
         Notes
@@ -146,9 +225,9 @@ class _Finder(_FinderBase):
 
         return glob.glob(pattern)
 
-    def _parse_paths(self, paths, on_parse_error):
+    def _parse_paths(self, paths, on_parse_error) -> pd.DataFrame:
 
-        out = list()
+        valid_paths, out = list(), list()
         for path in paths:
             parsed = self.parser.parse(path)
 
@@ -166,40 +245,40 @@ class _Finder(_FinderBase):
                 elif on_parse_error == "ignore":
                     pass
             else:
-                out.append([path + self._suffix] + list(parsed.named.values()))
+                valid_paths.append(path)
+                out.append(list(parsed.named.values()))
 
-        keys = ["filename"] + list(self.keys)
-
-        df = pd.DataFrame(out, columns=keys)
+        index = pd.Index(valid_paths, name="path") + self._suffix
+        df = pd.DataFrame(out, columns=self.keys, index=index)
         return df
 
 
 class FileFinder:
-    """find and create file names based on python format syntax
-
-    Parameters
-    ----------
-    path_pattern : str
-        String denoting the path (folder) pattern where everything variable is enclosed
-        in curly braces.
-    file_pattern : str
-        String denoting the file pattern where everything variable is enclosed in curly
-        braces.
-    test_paths : list of str, default None
-        A list of paths to use instead of querying the file system. To be used for
-        testing and demonstration.
-
-    Examples
-    --------
-    >>> path_pattern = "/root/{category}"
-    >>> file_pattern = "{category}_file_{number}"
-
-    >>> ff = FileFinder(path_pattern, file_pattern)
-    """
 
     def __init__(
         self, path_pattern: str, file_pattern: str, *, test_paths=None
     ) -> None:
+        """find and create file names based on python format syntax
+
+        Parameters
+        ----------
+        path_pattern : str
+            String denoting the path (folder) pattern where everything variable is
+            enclosed in curly braces.
+        file_pattern : str
+            String denoting the file pattern where everything variable is enclosed in
+            curly braces.
+        test_paths : list of str, default None
+            A list of paths to use instead of querying the file system. To be used for
+            testing and demonstration.
+
+        Examples
+        --------
+        >>> path_pattern = "/root/{category}"
+        >>> file_pattern = "{category}_file_{number}"
+
+        >>> ff = FileFinder(path_pattern, file_pattern)
+        """
 
         if os.path.sep in file_pattern:
             raise ValueError(
@@ -228,6 +307,8 @@ class FileFinder:
         if isinstance(test_paths, str):
             test_paths = [test_paths]
 
+        self._test_paths = test_paths
+
         # use fnmatch.filter to 'glob' pseudo-filenames
         def finder(pat):
 
@@ -243,7 +324,7 @@ class FileFinder:
         self.path._glob = finder
         self.full._glob = finder
 
-    def create_path_name(self, keys=None, **keys_kwargs):
+    def create_path_name(self, keys=None, **keys_kwargs) -> str:
         """build path (folder) name from keys
 
         Parameters
@@ -273,7 +354,7 @@ class FileFinder:
         # warnings.warn("'create_path_name' is deprecated, use 'path.name' instead")
         return self.path.create_name(keys, **keys_kwargs)
 
-    def create_file_name(self, keys=None, **keys_kwargs):
+    def create_file_name(self, keys=None, **keys_kwargs) -> str:
         """build file name from keys
 
         Parameters
@@ -303,7 +384,7 @@ class FileFinder:
         # warnings.warn("'create_file_name' is deprecated, use 'file.name' instead")
         return self.file.create_name(keys, **keys_kwargs)
 
-    def create_full_name(self, keys=None, **keys_kwargs):
+    def create_full_name(self, keys=None, **keys_kwargs) -> str:
         """build full (folder + file) name from keys
 
         Parameters
@@ -334,8 +415,8 @@ class FileFinder:
         return self.full.create_name(keys, **keys_kwargs)
 
     def find_paths(
-        self, keys=None, *, on_parse_error="raise", _allow_empty=False, **keys_kwargs
-    ):
+        self, keys=None, *, on_parse_error="raise", on_empty="raise", **keys_kwargs
+    ) -> "FileContainer":
         """find files in the file system using the file and path (folder) pattern
 
         Parameters
@@ -346,9 +427,9 @@ class FileFinder:
         on_parse_error : "raise" | "warn" | "skip", default: "raise"
             What to do if a path/file name cannot be parsed. If "raise" raises a ValueError,
             if "warn" raises a warning and if "skip" ignores the file.
-        _allow_empty : bool, default: False
-            If False (default) raises an error if no files are found. If True returns
-            an empty list.
+        on_empty : "raise" | "warn" | "allow", default: "raise"
+            Behaviour when no files are found: "raise" (default) raises a ValueError,
+            "warn" raises a warning. For "warn" and "allow" an empty FileContainer is returned.
         **keys_kwargs : {key: indexer, ...}, optional
             The keyword arguments form of ``keys``. When the same key is passed in
             ``keys`` and ``keys_kwargs`` the latter takes priority.
@@ -363,15 +444,15 @@ class FileFinder:
         >>> file_pattern = "{category}_file_{number}"
         >>> ff = FileFinder(path_pattern, file_pattern)
 
-        >>> ff.find()   # doctest: +SKIP
+        >>> ff.find_paths()   # doctest: +SKIP
         Looks for
         - "/root/*/"
 
-        >>> ff.find(category="foo")   # doctest: +SKIP
+        >>> ff.find_paths(category="foo")   # doctest: +SKIP
         Looks for
         - "/root/foo/"
 
-        >>> ff.find(dict(category=["foo", "bar"]))   # doctest: +SKIP
+        >>> ff.find_paths(dict(category=["foo", "bar"]))   # doctest: +SKIP
         Looks for
         - "/root/foo/"
         - "/root/bar/"
@@ -379,13 +460,13 @@ class FileFinder:
         return self.path.find(
             keys,
             on_parse_error=on_parse_error,
-            _allow_empty=_allow_empty,
+            on_empty=on_empty,
             **keys_kwargs,
         )
 
     def find_files(
-        self, keys=None, on_parse_error="raise", _allow_empty=False, **keys_kwargs
-    ):
+        self, keys=None, *, on_parse_error="raise", on_empty="raise", **keys_kwargs
+    ) -> "FileContainer":
         """find files in the file system using the file pattern
 
         Parameters
@@ -396,9 +477,9 @@ class FileFinder:
         on_parse_error : "raise" | "warn" | "skip", default: "raise"
             What to do if a path/file name cannot be parsed. If "raise" raises a ValueError,
             if "warn" raises a warning and if "skip" ignores the file.
-        _allow_empty : bool, default: False
-            If False (default) raises an error if no files are found. If True returns
-            an empty list.
+        on_empty : "raise" | "warn" | "allow", default: "raise"
+            Behaviour when no files are found: "raise" (default) raises a ValueError,
+            "warn" raises a warning. For "warn" and "allow" an empty FileContainer is returned.
         **keys_kwargs : {key: indexer, ...}, optional
             The keyword arguments form of ``keys``. When the same key is passed in
             ``keys`` and ``keys_kwargs`` the latter takes priority.
@@ -432,11 +513,57 @@ class FileFinder:
         return self.full.find(
             keys,
             on_parse_error=on_parse_error,
-            _allow_empty=_allow_empty,
+            on_empty=on_empty,
             **keys_kwargs,
         )
 
-    def __repr__(self):
+    def find_single_path(self, keys=None, **keys_kwargs) -> "FileContainer":
+        """
+        find exactly one path in the file system using the path pattern
+
+        Parameters
+        ----------
+        keys : dict
+            Dictionary containing keys to create the search pattern.
+        **keys_kwargs : {key: indexer, ...}, optional
+            The keyword arguments form of ``keys``. When the same key is passed in
+            ``keys`` and ``keys_kwargs`` the latter takes priority.
+
+        Notes
+        -----
+        Missing ``keys`` are replaced with ``"*"``.
+
+        Raises
+        ------
+        ValueError : if more or less than one path is found
+        """
+
+        return self.path.find_single(keys, **keys_kwargs)
+
+    def find_single_file(self, keys=None, **keys_kwargs) -> "FileContainer":
+        """
+        find exactly one file in the file system using the file and path pattern
+
+        Parameters
+        ----------
+        keys : dict
+            Dictionary containing keys to create the search pattern.
+        **keys_kwargs : {key: indexer, ...}, optional
+            The keyword arguments form of ``keys``. When the same key is passed in
+            ``keys`` and ``keys_kwargs`` the latter takes priority.
+
+        Notes
+        -----
+        Missing ``keys`` are replaced with ``"*"``.
+
+        Raises
+        ------
+        ValueError : if more or less than one file is found
+        """
+
+        return self.full.find_single(keys, **keys_kwargs)
+
+    def __repr__(self) -> str:
 
         repr_keys = "', '".join(sorted(self.full.keys))
         repr_keys = f"'{repr_keys}'"
@@ -451,16 +578,22 @@ class FileFinder:
 
 
 class FileContainer:
-    """docstring for FileContainer"""
 
-    def __init__(self, df):
+    def __init__(self, df: pd.DataFrame):
+        """FileContainer gathers paths and their metadata
+
+        Parameters
+        ----------
+        df : pd.DataFrame
+            DataFrame with info about found paths from FileFinder.
+        """
 
         self.df = df
 
     def __iter__(self):
 
         for index, element in self.df.iterrows():
-            yield element["filename"], element.drop("filename").to_dict()
+            yield index, element.to_dict()
 
     def __getitem__(self, key):
 
@@ -468,26 +601,55 @@ class FileContainer:
             # use iloc -> there can be more than one element with index 0
             element = self.df.iloc[key]
 
-            return element["filename"], element.drop("filename").to_dict()
+            return element.name, element.to_dict()
         # assume slice or [1]
         else:
             ret = copy.copy(self)
             ret.df = self.df.iloc[key]
             return ret
 
+    @property
+    def meta(self) -> list[dict[str, Any]]:
+        return self.df.to_dict("records")
+
+    @property
+    def paths(self) -> list[str]:
+        return self.df.index.to_list()
+
     def combine_by_key(self, keys=None, sep="."):
+        warnings.warn(
+            "`combine_by_key` has been deprecated and will be removed in a future version",
+            FutureWarning,
+        )
+
+        return self._combine_by_keys(keys=keys, sep=sep)
+
+    def _combine_by_keys(self, keys=None, sep="."):
         """combine columns"""
 
         if keys is None:
-            keys = list(self.df.columns.drop("filename"))
+            keys = list(self.df.columns)
 
         return self.df[list(keys)].apply(lambda x: sep.join(x.map(str)), axis=1)
 
     def search(self, **query):
+        """subset paths given a search query
 
-        ret = copy.copy(self)
-        ret.df = self._get_subset(**query)
-        return ret
+        Parameters
+        ----------
+        **query: Mapping[str, str | int | list[str | int]]
+            Search query.
+
+        Notes
+        -----
+        - individual conditions are combined with "and", e.g., ``model="a", exp="b"``
+          requires the model to be "a" and the experiment to be "b".
+        - conditions for a key are combined with "or", e.g., ``model=["a", "b"]``
+          matches for both.
+        """
+
+        df = self._get_subset(**query)
+        return type(self)(df)
     
     def concat(self, other):
         """concatenate two FileContainers"""
@@ -504,23 +666,25 @@ class FileContainer:
 
     def _get_subset(self, **query):
         if not query:
-            return pd.DataFrame(columns=self.df.columns)
-        condition = np.ones(len(self.df), dtype=bool)
-        for key, val in query.items():
-            if isinstance(val, list):
-                condition_i = np.zeros(len(self.df), dtype=bool)
-                for val_i in val:
-                    condition_i = condition_i | (self.df[key] == val_i)
-                condition = condition & condition_i
-            elif val is not None:
-                condition = condition & (self.df[key] == val)
-        query_results = self.df.loc[condition]
-        return query_results
+            return pd.DataFrame(
+                [], columns=self.df.columns, index=pd.Index([], name="path")
+            )
+
+        sel = True
+        for key, value in query.items():
+            # isin does not handle scalars
+            value = [value] if np.ndim(value) == 0 else value
+
+            sel &= self.df[key].isin(value)
+
+        return self.df[sel]
 
     def __len__(self):
         return self.df.__len__()
 
     def __repr__(self):
 
-        msg = "<FileContainer>\n"
+        n_paths = len(self)
+
+        msg = f"<FileContainer: {n_paths} paths>\n"
         return msg + self.df.__repr__()