+++ filefinder/tests/test_filefinder.py
@@ -5,6 +5,9 @@ import pandas as pd
 import pytest
 
 from filefinder import FileFinder
+from filefinder._filefinder import _assert_unique
+
+from . import assert_filecontainer_empty
 
 
 @pytest.fixture(scope="module")
@@ -45,6 +48,60 @@ def test_paths(request, tmp_path):
     return paths
 
 
+@pytest.mark.parametrize("placeholder", ("keys", "on_parse_error", "_allow_empty"))
+def test_pattern_invalid_placeholder(placeholder):
+
+    with pytest.raises(ValueError, match=f"'{placeholder}' is not a valid placeholder"):
+        FileFinder("", f"{{{placeholder}}}")
+
+    with pytest.raises(ValueError, match=f"'{placeholder}' is not a valid placeholder"):
+        FileFinder(f"{{{placeholder}}}", "")
+
+
+@pytest.mark.parametrize("pattern", ("{}", "{_fixed}"))
+def test_only_named_fields(pattern):
+
+    with pytest.raises(ValueError, match="Only named fields are currently allowed"):
+        FileFinder("", pattern)
+
+    with pytest.raises(ValueError, match="Only named fields are currently allowed"):
+        FileFinder(pattern, "")
+
+
+def test_assert_unique():
+
+    # no error raised
+    df = pd.DataFrame.from_records([("a", "d"), ("a", "h")], columns=("model", "res"))
+    _assert_unique(df)
+
+    df = pd.DataFrame.from_records([("a", "d"), ("a", "d")], columns=("model", "res"))
+    with pytest.raises(ValueError, match="Non-unique metadata detected"):
+        _assert_unique(df)
+
+
+@pytest.mark.parametrize("_allow_empty", (False, True))
+def test_deprecate_allow_empty(_allow_empty):
+
+    ff = FileFinder("", "a")
+    msg = "`_allow_empty` has been deprecated in favour of `on_empty`"
+    with pytest.raises(TypeError, match=msg):
+        ff.find_files(_allow_empty=_allow_empty)
+
+    with pytest.raises(TypeError, match=msg):
+        ff.find_paths(_allow_empty=_allow_empty)
+
+
+def test_wrong_on_empty():
+
+    ff = FileFinder("", "a")
+    msg = "Unknown value for 'on_empty': 'null'. Must be one of 'raise', 'warn' or 'allow'."
+    with pytest.raises(ValueError, match=msg):
+        ff.find_paths(on_empty="null")
+
+    with pytest.raises(ValueError, match=msg):
+        ff.find_files(on_empty="null")
+
+
 def test_pattern_property():
 
     path_pattern = "path_pattern/"
@@ -61,6 +118,20 @@ def test_pattern_property():
     assert ff.file.pattern == file_pattern
 
     assert ff.full.pattern == path_pattern + file_pattern
+
+
+def test_test_path_property():
+
+    ff = FileFinder("a", "b")
+
+    with pytest.raises(AttributeError):
+        ff._test_paths
+
+    ff = FileFinder("a", "b", test_paths="path")
+    assert ff._test_paths == ["path"]
+
+    ff = FileFinder("a", "b", test_paths=["a", "b"])
+    assert ff._test_paths == ["a", "b"]
 
 
 def test_file_pattern_no_sep():
@@ -191,11 +262,29 @@ def test_find_path_none_found(tmp_path, test_paths):
     with pytest.raises(ValueError, match="Found no files matching criteria"):
         ff.find_paths({"a": "foo"})
 
-    result = ff.find_paths(a="foo", _allow_empty=True)
-    assert result == []
+    with pytest.warns(match="Found no files matching criteria"):
+        result = ff.find_paths(a="foo", on_empty="warn")
+    assert_filecontainer_empty(result, columns="a")
 
-    result = ff.find_paths({"a": "foo"}, _allow_empty=True)
-    assert result == []
+    with pytest.warns(match="Found no files matching criteria"):
+        result = ff.find_paths({"a": "foo"}, on_empty="warn")
+    assert_filecontainer_empty(result, columns="a")
+
+    result = ff.find_paths(a="foo", on_empty="allow")
+    assert_filecontainer_empty(result, columns="a")
+
+    result = ff.find_paths({"a": "foo"}, on_empty="allow")
+    assert_filecontainer_empty(result, columns="a")
+
+
+def test_find_paths_non_unique():
+
+    # test raises error for non-unique metadata - AFAIK not possible for real paths
+
+    ff = FileFinder("{cat}", "", test_paths=["a/", "a/"])
+
+    with pytest.raises(ValueError, match="Non-unique metadata detected"):
+        ff.find_paths()
 
 
 def test_find_paths_simple(tmp_path, test_paths):
@@ -207,8 +296,8 @@ def test_find_paths_simple(tmp_path, test_paths):
         path_pattern=path_pattern, file_pattern=file_pattern, test_paths=test_paths
     )
 
-    expected = {"filename": {0: str(tmp_path / "a1/foo/*")}, "a": {0: "foo"}}
-    expected = pd.DataFrame.from_dict(expected)
+    expected = {"path": {0: str(tmp_path / "a1/foo/*")}, "a": {0: "foo"}}
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_paths(a="foo")
     pd.testing.assert_frame_equal(result.df, expected)
@@ -231,11 +320,11 @@ def test_find_paths_wildcard(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {0: str(tmp_path / "a1/foo/*"), 1: str(tmp_path / "a2/foo/*")},
+        "path": {0: str(tmp_path / "a1/foo/*"), 1: str(tmp_path / "a2/foo/*")},
         "a": {0: "a1", 1: "a2"},
         "b": {0: "foo", 1: "foo"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_paths(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -261,11 +350,11 @@ def test_find_paths_several(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {0: str(tmp_path / "a1/foo/*"), 1: str(tmp_path / "a2/foo/*")},
+        "path": {0: str(tmp_path / "a1/foo/*"), 1: str(tmp_path / "a2/foo/*")},
         "a": {0: "a1", 1: "a2"},
         "b": {0: "foo", 1: "foo"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_paths(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -291,11 +380,11 @@ def test_find_paths_one_of_several(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {0: str(tmp_path / "a1/foo/*")},
+        "path": {0: str(tmp_path / "a1/foo/*")},
         "a": {0: "a1"},
         "b": {0: "foo"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_paths(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -304,6 +393,33 @@ def test_find_paths_one_of_several(tmp_path, test_paths, find_kwargs):
     pd.testing.assert_frame_equal(result.df, expected)
 
     result = ff.find_paths({"a": "XXX"}, **find_kwargs)
+    pd.testing.assert_frame_equal(result.df, expected)
+
+
+def test_find_single_path(tmp_path, test_paths):
+
+    path_pattern = tmp_path / "{a}/foo"
+    file_pattern = "file_pattern"
+
+    ff = FileFinder(
+        path_pattern=path_pattern, file_pattern=file_pattern, test_paths=test_paths
+    )
+
+    # error if more than one is found
+    with pytest.raises(ValueError, match=r"Found more than one \(2\) files/ paths"):
+        ff.find_single_path()
+
+    # error if more than one is found
+    with pytest.raises(ValueError, match="Found no files matching criteria"):
+        ff.find_single_path(a="a3")
+
+    expected = {"path": {0: str(tmp_path / "a1/foo/*")}, "a": {0: "a1"}}
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
+
+    result = ff.find_single_path(a="a1")
+    pd.testing.assert_frame_equal(result.df, expected)
+
+    result = ff.find_single_path({"a": "a1"})
     pd.testing.assert_frame_equal(result.df, expected)
 
 
@@ -322,14 +438,32 @@ def test_find_file_none_found(tmp_path, test_paths):
     with pytest.raises(ValueError, match="Found no files matching criteria"):
         ff.find_files({"a": "XXX"})
 
-    result = ff.find_files(a="XXX", _allow_empty=True)
-    assert result == []
+    with pytest.warns(match="Found no files matching criteria"):
+        result = ff.find_files(a="XXX", on_empty="warn")
+    assert_filecontainer_empty(result, columns=("a", "file_pattern"))
 
-    result = ff.find_files({"a": "XXX"}, _allow_empty=True)
-    assert result == []
+    with pytest.warns(match="Found no files matching criteria"):
+        result = ff.find_files({"a": "XXX"}, on_empty="warn")
+    assert_filecontainer_empty(result, columns=("a", "file_pattern"))
 
-    result = ff.find_files({"a": "XXX"}, _allow_empty=True, a="XXX")
-    assert result == []
+    result = ff.find_files(a="XXX", on_empty="allow")
+    assert_filecontainer_empty(result, columns=("a", "file_pattern"))
+
+    result = ff.find_files({"a": "XXX"}, on_empty="allow")
+    assert_filecontainer_empty(result, columns=("a", "file_pattern"))
+
+    result = ff.find_files({"a": "XXX"}, on_empty="allow", a="XXX")
+    assert_filecontainer_empty(result, columns=("a", "file_pattern"))
+
+
+def test_find_files_non_unique():
+
+    # test raises error for non-unique metadata - AFAIK not possible for real paths
+
+    ff = FileFinder("", "{cat}", test_paths=["/a", "/a"])
+
+    with pytest.raises(ValueError, match="Non-unique metadata detected"):
+        ff.find_files()
 
 
 def test_find_file_simple(tmp_path, test_paths):
@@ -341,8 +475,8 @@ def test_find_file_simple(tmp_path, test_paths):
         path_pattern=path_pattern, file_pattern=file_pattern, test_paths=test_paths
     )
 
-    expected = {"filename": {0: str(tmp_path / "a1/foo/file")}, "a": {0: "foo"}}
-    expected = pd.DataFrame.from_dict(expected)
+    expected = {"path": {0: str(tmp_path / "a1/foo/file")}, "a": {0: "foo"}}
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_files(a="foo")
     pd.testing.assert_frame_equal(result.df, expected)
@@ -365,14 +499,14 @@ def test_find_files_wildcard(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {
+        "path": {
             0: str(tmp_path / "a1/foo/file"),
             1: str(tmp_path / "a2/foo/file"),
         },
         "a": {0: "a1", 1: "a2"},
         "b": {0: "file", 1: "file"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_files(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -398,14 +532,14 @@ def test_find_files_several(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {
+        "path": {
             0: str(tmp_path / "a1/foo/file"),
             1: str(tmp_path / "a2/foo/file"),
         },
         "a": {0: "a1", 1: "a2"},
         "b": {0: "file", 1: "file"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_files(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -431,11 +565,11 @@ def test_find_files_one_of_several(tmp_path, test_paths, find_kwargs):
     )
 
     expected = {
-        "filename": {0: str(tmp_path / "a1/foo/file")},
+        "path": {0: str(tmp_path / "a1/foo/file")},
         "a": {0: "a1"},
         "b": {0: "file"},
     }
-    expected = pd.DataFrame.from_dict(expected)
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
 
     result = ff.find_files(**find_kwargs)
     pd.testing.assert_frame_equal(result.df, expected)
@@ -447,14 +581,41 @@ def test_find_files_one_of_several(tmp_path, test_paths, find_kwargs):
     pd.testing.assert_frame_equal(result.df, expected)
 
 
+def test_find_single_file(tmp_path, test_paths):
+
+    path_pattern = tmp_path / "{a}/foo"
+    file_pattern = "file"
+
+    ff = FileFinder(
+        path_pattern=path_pattern, file_pattern=file_pattern, test_paths=test_paths
+    )
+
+    # error if more than one is found
+    with pytest.raises(ValueError, match=r"Found more than one \(2\) files/ paths"):
+        ff.find_single_file()
+
+    # error if more than one is found
+    with pytest.raises(ValueError, match="Found no files matching criteria"):
+        ff.find_single_file(a="a3")
+
+    expected = {"path": {0: str(tmp_path / "a1/foo/file")}, "a": {0: "a1"}}
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
+
+    result = ff.find_single_file(a="a1")
+    pd.testing.assert_frame_equal(result.df, expected)
+
+    result = ff.find_single_file({"a": "a1"})
+    pd.testing.assert_frame_equal(result.df, expected)
+
+
 def test_find_paths_scalar_number():
 
     ff = FileFinder(
         path_pattern="{path}", file_pattern="{file}", test_paths=["1/1", "2/2"]
     )
 
-    expected = {"filename": {0: "1/*"}, "path": {0: "1"}}
-    expected = pd.DataFrame.from_dict(expected)
+    index = pd.Index(["1/*"], name="path")
+    expected = pd.DataFrame(["1"], columns=["path"], index=index)
     result = ff.find_paths(path=1)
     pd.testing.assert_frame_equal(result.df, expected)
 
@@ -465,8 +626,8 @@ def test_find_files_scalar_number():
         path_pattern="{path}", file_pattern="{file}", test_paths=["1/1", "2/2"]
     )
 
-    expected = {"filename": {0: "1/1"}, "path": {0: "1"}, "file": {0: "1"}}
-    expected = pd.DataFrame.from_dict(expected)
+    index = pd.Index(["1/1"], name="path")
+    expected = pd.DataFrame([["1", "1"]], columns=["path", "file"], index=index)
     result = ff.find_files(file=1)
     pd.testing.assert_frame_equal(result.df, expected)
 
@@ -480,7 +641,7 @@ def test_find_unparsable():
     ):
         ff.find_files()
 
-    expected = pd.DataFrame(list(), columns=["filename", "cat"])
+    expected = pd.DataFrame([], columns=["cat"], index=pd.Index([], name="path"))
 
     with pytest.warns(match="Could not parse 'a/b' with the pattern '{cat}/{cat}'"):
         result = ff.find_files(on_parse_error="warn")
@@ -490,8 +651,8 @@ def test_find_unparsable():
     pd.testing.assert_frame_equal(result.df, expected)
 
     ff = FileFinder("{cat}", "{cat}", test_paths=["a/b", "a/a"])
-    expected = {"filename": {0: "a/a"}, "cat": {0: "a"}}
-    expected = pd.DataFrame.from_dict(expected)
+    expected = {"path": {0: "a/a"}, "cat": {0: "a"}}
+    expected = pd.DataFrame.from_dict(expected).set_index("path")
     result = ff.find_files(on_parse_error="ignore")
     pd.testing.assert_frame_equal(result.df, expected)